add dedicated extensions/wlroots file
