+++ src/extensions/wayland.h
@@ -2,7 +2,6 @@
 #define CATNIP_EXTENSIONS_WAYLAND_H
 
 #include <wayland-server-core.h>
-#include <wlr/types/wlr_scene.h>
 
 void
 wl_signal_subscribe(
@@ -10,11 +9,5 @@ wl_signal_subscribe(
   struct wl_listener* listener,
   wl_notify_func_t callback
 );
-
-int
-wlr_scene_node_get_zindex(struct wlr_scene_node* node);
-
-void
-wlr_scene_node_set_zindex(struct wlr_scene_node* node, int zindex);
 
 #endif
